% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BackTestingRealTime.R
\name{BackTestingRealTime}
\alias{BackTestingRealTime}
\title{Real Time Backtesting Function for Modeling}
\usage{
BackTestingRealTime(context, datasets, nEval = 500, longTrigger = -1,
  shortTrigger = 1, exitTrigger = 0, strategyFunc = NULL,
  strategyParam = 5)
}
\arguments{
\item{context}{trading context}

\item{datasets}{a dataset to be used for modeling}

\item{nEval}{length of periods used for training}

\item{longTrigger}{a number of thresholds for going long}

\item{shortTrigger}{a number of thresholds for going short}

\item{exitTrigger}{a trigger to exit current positions}

\item{strategyFunc}{customised function to calculate indicator from \code{datasets}}

\item{strategyParam}{parameters used in \code{strategyFunc}}
}
\value{
A \code{data.frame} with trading activities
}
\description{
Real time backtesting function for modeling with removed survivor biased and bias of using future elements
}
\seealso{
\link{getUserTemplate}
}

