% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JohansenCointegrationTest.R
\name{findCointegratedPairs}
\alias{findCointegratedPairs}
\title{Search Cointegrated Pairs}
\usage{
findCointegratedPairs(data)
}
\arguments{
\item{data}{Data matrix including all serires}
}
\description{
Conducts the Johansen procedure and ADF test on all combination of pairs to find potential integrated pairs
}
\seealso{
\link{JohansenCointegrationTest}
}

