% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUserTemplate.R
\name{getUserTemplate}
\alias{getUserTemplate}
\title{Get a Template Run of AutoPairTrading Package}
\usage{
getUserTemplate()
}
\description{
Get a Template Run of AutoPairTrading Package
}
\examples{
getUserTemplate()

}

