% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUserTemplate.R
\name{getUserTemplate2}
\alias{getUserTemplate2}
\title{Get a Template Run of AutoPairTrading Package}
\usage{
getUserTemplate2()
}
\description{
Get a Template Run of AutoPairTrading Package
}
\examples{
getUserTemplate2()

}

