% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HedgeRatioOLS.R
\name{HedgeRatioOLS}
\alias{HedgeRatioOLS}
\title{Hedge Ratio OLS}
\usage{
HedgeRatioOLS(y, x, add_const = TRUE)
}
\arguments{
\item{y}{long series}

\item{x}{short series}

\item{add_const}{keep constant (risk free profits) or not}
}
\value{
A \code{list} of hedge ratio (beta, risk) and alpha (risk free profits)
}
\description{
Calculate Hedge Ratio by using OLS
}
\examples{
getFX("AUD/USD")
getFX("CAD/USD")
HedgeRatioOLS(AUDUSD, CADUSD)

}

