% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HurstExponentTest.R
\name{HurstExponentTest}
\alias{HurstExponentTest}
\title{Hurst Exponent Test}
\usage{
HurstExponentTest(y, type, lags)
}
\arguments{
\item{y}{Vector to be tested for a unit root.}

\item{type}{Test type, either "none", "drift" or "trend".}

\item{lags}{Number of lags for endogenous variable to be included.}
}
\value{
A \code{list} of ADF test results
}
\description{
Effectively this returns a value between 0 and 1 that tells you whether a time-series is trending or mean-reverting.
The closer the value is to 0.5 means the more "random" the time-series has behaved historically.
Values below 0.5 imply the time-series is mean-reverting, and above 0.5 imply trending.
The closer the value is to 0 implies greater levels of mean-reversion.
}
\examples{
data(Raotbl3)
attach(Raotbl3)
lc.df <- AugmentedDickeyFullerTest(y=lc, lags=3, type='trend')
summary(lc.df)

}

