% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitializeContext.R
\name{InitializeContext}
\alias{InitializeContext}
\title{Initialize Trading Context}
\usage{
InitializeContext(Symbol.Y, Symbol.X, capital = 1e+05, window = 20,
  lookback = 250, brokerage = 0.02, stoploss = 0.1, half.life)
}
\arguments{
\item{Symbol.Y}{Ticker symbol for series Y (long)}

\item{Symbol.X}{Ticker symbol for series X (short)}

\item{capital}{Initial capital for investment}

\item{window}{used for zscore calculation, must be <= \code{lookback}}

\item{lookback}{used for regression}

\item{brokerage}{Commission fees in percentage}

\item{stoploss}{Max drawdown for stop loss in percentage}
}
\value{
A \code{list} of attributes about current trading context
}
\description{
Initialze and configure trading context
}
\examples{
getFX("AUD/USD")
getFX("CAD/USD")
context <- InitializeContext(AUDUSD, CADUSD, capital = 1e5, window = 20, lookback = 250, brokerage = 0.02, stoploss = 0.1)
context

}

