% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JohansenCointegrationTest.R
\name{JohansenCointegrationTest}
\alias{JohansenCointegrationTest}
\title{Johansen Cointegration Test}
\usage{
JohansenCointegrationTest(y, type, lags)
}
\arguments{
\item{y}{Vector to be tested for a unit root.}

\item{type}{Test type, either "none", "drift" or "trend".}

\item{lags}{Number of lags for endogenous variable to be included.}
}
\value{
A \code{list} of ADF test results
}
\description{
Conducts the Johansen procedure on a given data set.
The "trace" or "eigen" statistics are reported and the matrix of eigenvectors as well as the loading matrix.
}
\examples{
data(Raotbl3)
attach(Raotbl3)
lc.df <- AugmentedDickeyFullerTest(y=lc, lags=3, type='trend')
summary(lc.df)

}

