% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformanceAnalytics.R
\name{performanceReport}
\alias{performanceReport}
\title{Get a basic report of trading strategy}
\usage{
performanceReport(dt.summary)
}
\arguments{
\item{dt.summary}{trading activity sheet returned from \code{BackTesting}}
}
\value{
A \code{list} of basic trading details
}
\description{
Get a basic report of trading strategy
}
\seealso{
\link{BackTesting}
}

