% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantstratOrders.R
\name{put.orderbook}
\alias{put.orderbook}
\title{put an orderbook object in .strategy env}
\usage{
put.orderbook(portfolio.st, orderbook, envir = .strategyEnv)
}
\arguments{
\item{portfolio.st}{string identifying portfolio}

\item{orderbook}{orderbook object}

\item{envir}{the environment to store the orderbook object in, defaults to .strategy}
}
\description{
put an orderbook object in .strategy env
}
\seealso{
getOrderBook
}
\concept{
order book
}

