% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseConnection.R
\name{QuandlConnect}
\alias{QuandlConnect}
\title{Quandl API Connection}
\usage{
QuandlConnect()
}
\description{
Connect to Quandl API
}
\examples{
QuandlConnect()
mydata = Quandl.datatable("ZACKS/FC", ticker="AAPL")

}

