% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchCointegrationPairs.R
\name{searchCointegratedPairs}
\alias{searchCointegratedPairs}
\title{Search Cointegrated Pairs}
\usage{
searchCointegratedPairs(dataset, path = "GoodIntegratedPairs.pdf",
  to = c("ivan.liuyanfeng@gmail.com"), testPeriod = 63, trainPeriod = 252)
}
\arguments{
\item{dataset}{Data matrix including all serires}

\item{path}{the path of the pdf report to be saved}

\item{to}{a list of email addressed to receive the reports}

\item{testPeriod}{number of points used to test cointegrations}

\item{trainPeriod}{number of points used to train cointegrations}
}
\description{
Conduct the Johansen procedure and ADF test on all combination of pairs to find potential integrated pairs
Conduct the out of sample tests on all pairs to test the consistencies of cointegrations
}
\seealso{
\link{JohansenCointegrationTest}
}

