% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TradingStrategyFunctions.R
\name{strategyBuilder}
\alias{strategyBuilder}
\title{Generate Indicators of Entry/Exit points}
\usage{
strategyBuilder(datasets, half.life = 10, itr = 10)
}
\arguments{
\item{datasets}{the latest datasets for modeling (first columns should always be the target series. e.g. price ratio)}

\item{half.life}{half-life of mean reversion series}

\item{itr}{default 10, number of iterations for simulations, the larger the better performance and longer running time}
}
\value{
An indicator of entry/exit
}
\description{
Use rule based and machine learning predictions to generate indicators for determining next trading activities.
}

