% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MachineLearningSimulation.R
\name{xgbRegSimulation}
\alias{xgbRegSimulation}
\title{Xgboost Simulations for Regression}
\usage{
xgbRegSimulation(idx, features, labels, pred, itr = 10)
}
\arguments{
\item{idx}{index for train/test split}

\item{features}{feature data used for training}

\item{labels}{objective variables}

\item{pred}{today's features for predicting}

\item{itr}{default 10, number of iterations for simulations, the larger the better performance and longer running time}
}
\value{
A vector of estimated future slopes/movements
}
\description{
Use xgboost algorithm to predict future movements of the price ratios
}

