% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BackTesting.R
\name{BackTesting}
\alias{BackTesting}
\title{Back Testing Strategy}
\usage{
BackTesting(y, x, context, strategy, goLong = -1, goShort = 1)
}
\arguments{
\item{y}{long series}

\item{x}{short series}

\item{context}{trading context}

\item{strategy}{a vector of trading indicator calculated by user's strategy}

\item{goLong}{threshold for going long}

\item{goShort}{threshold for going short}
}
\value{
A \code{data.frame} with trading activities
}
\description{
Back Testing Strategy
}

