% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HedgeRatioOLS.R
\name{HedgeRatioOLS}
\alias{HedgeRatioOLS}
\title{Hedge Ratio OLS}
\usage{
HedgeRatioOLS(y, x, add_const = TRUE)
}
\arguments{
\item{y}{long series}

\item{x}{short series}

\item{add_const}{keep constant (risk free profits) or not}
}
\value{
A \code{list} of hedge ratio (beta, risk) and alpha (risk free profits)
}
\description{
Calculate Hedge Ratio by using OLS
}
\examples{
data(Raotbl3)
attach(Raotbl3)
lc.df <- AugmentedDickeyFullerTest(y=lc, lags=3, type='trend')
summary(lc.df)

}

